:hammer: Rename opt. proc to optim_recipe
