+++ auto_tutorials_source/tutorial_bayesian.py
@@ -105,7 +105,7 @@ routine = ClassificationRoutine(
     model=model,
     num_classes=datamodule.num_classes,
     loss=loss,
-    optimization_procedure=optim_lenet,
+    optim_recipe=optim_lenet,
 )
 
 # %%
