+++ auto_tutorials_source/tutorial_mc_batch_norm.py
@@ -19,7 +19,7 @@ First, we have to load the following utilities from TorchUncertainty:
 - the mc-batch-norm wrapper: mc_dropout, which lies in torch_uncertainty.models
 - a resnet baseline to get the command line arguments: ResNet, which lies in torch_uncertainty.baselines
 - the classification training routine in the torch_uncertainty.training.classification module
-- the optimizer wrapper in the torch_uncertainty.optimization_procedures module.
+- the optimizer wrapper in the torch_uncertainty.optim_recipes module.
 
 We also need import the neural network utils withing `torch.nn`.
 """
@@ -31,7 +31,7 @@ from torch import nn
 
 from torch_uncertainty.datamodules import MNISTDataModule
 from torch_uncertainty.models.lenet import lenet
-from torch_uncertainty.optimization_procedures import optim_cifar10_resnet18
+from torch_uncertainty.optim_recipes import optim_cifar10_resnet18
 from torch_uncertainty.post_processing.mc_batch_norm import MCBatchNorm
 from torch_uncertainty.routines import ClassificationRoutine
 
@@ -67,7 +67,7 @@ routine = ClassificationRoutine(
     num_classes=datamodule.num_classes,
     model=model,
     loss=nn.CrossEntropyLoss,
-    optimization_procedure=optim_cifar10_resnet18,
+    optim_recipe=optim_cifar10_resnet18,
 )
 
 # %%
