+++ tests/baselines/test_mimo.py
@@ -4,7 +4,7 @@ from torchinfo import summary
 
 from torch_uncertainty.baselines.classification import ResNet, WideResNet
 
-# from torch_uncertainty.optimization_procedures import (
+# from torch_uncertainty.optim_recipes import (
 #     optim_cifar10_resnet18,
 #     optim_cifar10_resnet50,
 #     optim_cifar10_wideresnet,
