+++ tests/test_optimization_procedures.py
@@ -4,7 +4,7 @@ import pytest
 from torch_uncertainty.models.resnet import resnet18, resnet34, resnet50
 from torch_uncertainty.models.vgg import vgg16
 from torch_uncertainty.models.wideresnet import wideresnet28x10
-from torch_uncertainty.optimization_procedures import (
+from torch_uncertainty.optim_recipes import (
     get_procedure,
     optim_regression,
 )
